Implement support for the AVX2 compile time option
Currently this adds a complete new set of build steps with different flags,
with the aim of being placed in /usr/lib64/avx2. Our glibc, where appropriate,
will automatically load these libraries instead of the normal libraries inside
/usr/lib64/.

Right now we're simply setting the march/mtune to Haswell, which is the minimum
baseline required for AVX2 instruction set, but *may* expand or adjust this in
future for compatibility with non-Intel CPUs, depending on how portable and
feasible it is.

The sole aim here is to provide further optimisations and offer seamless use
of "better" libraries so that users can automatically benefit from libraries
that are optimizes for their CPUs, as opposed to being restricted to the current
stock offering of "-march=x86-64 -mtune=generic".

The original concept of using the separate avx2 directory and implementing a
TLS-like provider in the ld cascade came from the Clear Linux Project, and
has been documented here: https://dev.solus-project.com/T503

The work added in ypkg to support this is native to ypkg and original within
this tree.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>