examine: Support multiple directories in the RPATH
As evil as the rpath is, it's infinitely more so when it comes to the
ghc Haskell packages. These use a split list of directories for the rpath,
creating much work for the dynamic linker...

This change ensures we can at least find and support the dependendent libs
throughout these rpaths and not encounter missing symbols at any time.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>