ypkg-gen-history: Check every word to determine CVE status
Currently, we're not marking all packages as security updates even when they
explicitly are. This is because we match the entire block and don't match on
the words. This is an oversight and gives the impression that less security
updates are happening within Solus than actually are happening, such as the
recent libx11 libs being updated and not being displayed in the security
center as actual CVE fixes.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>