Finish of the AVX2 implementation
With this change we now build AVX2 libraries for both 32-bit and 64-bit
builds, and do not make them dependent on a particular base CPU. The
base level is determined by the CPU features only.

With the glibc changes in Solus, adapted from Clear Linux, these avx2
libraries will only be loaded when glibc detects that it is ok to do so.
Another minor change was to stop using /emul32 prefix on 32-bit builds,
as they are always done first they always get overriden.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>