spec: Officially support the 'networking' key
This key is in fact only used by solbuild, however it is part of the ypkg
specification. By default, this key is disabled, which means that sobuild
will in turn disable networking in the container isolation of the build.

During migration, and perhaps even for a few corner cases, it might still
be desirable to have networking support during the build. In this instance
the packager need only set the key to true.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>