dependencies: Improve dependency resolution by a ridiculous amount
We shouldn't be trying to search_file the dependency, as we already
know the exact file paths that we're intested in. Using search_file
will open and regexp scan the file infos and check every one of them,
every time, leading to potentially thousands of opens and regexp
compiles.

This method checks for the hashed key in the filesdb and just returns
the package name if we're known to have it, otherwise it'll return nil.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>