examine: Inspect all of the files in parallel
This change restores parallel processing to the PackageExaminer, which
should help to significantly reduce the amount of time spent looking at
the file set we've been handed.

As we're liable for race conditions in shared directory names, we first
assume that makedirs will fail, as another thread may have already
created the directory. Thus the error is fallthrough, and we check
again immediately for it's existence.

This will help creation of dbginfo packages and still retain a highly
parallel examination routine.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>