Add support for permanent paths in packages
As per issue #23, this change adds support for permanent paths to ypkg.
These files are matched by the built-in ypkg pattern system, and any
resulting path is then marked as being "permanent" within the resulting
.eopkg's.

These packages contain "resident" files, i.e. files that are *not* removed
on upgrade. Normal removals will cause the file to "disappear", but the
files will stay on the filesystem in the event of an upgrade.

This should be used sparingly, and only in areas where the domain of
control lies outside of the update delivery mechanism, i.e. kernel and
driver packages.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>