examine: Ensure we account for private dependencies in pkgconfig scan
We sometimes end up missing our pkgconfig dependencies due to private
requires not being met in the pkgconfig calculations. To mitigate this
we ensure pkg-config knows about all potential pkgconfigs within the
package as well as the appropriate host-side pkgconfigs, so that full
pkg-config dependencies are mapped out for cases such as mutter, which
has private-dep requirements.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>