dependencies: Ensure 32-bit and 64-bit sonames are tracked independently
This fixes an issue with the autogen build, which will perform a soname lookup
to find out the library that autogen depends on (libopt.so.25) - which was
accidentally coming from autogen-32bit, and not the main package. This further
became visible when splitting autogen-libs to replace libautogen, adding an
invalid dependency because we didn't separate 32-bit from 64-bit soname
providers, allowing crossed dependencies.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>