packages: Add pattern to move KDE dev docs to devel
QCH files are used with qt-creator and kdevelop, which are helpful only for
people developing on Solus. If they are building using a lib, they will need
the -devel package installed so they will also get these files. Want to move
them out of "main" as they can be significantly larger than the smaller
libraries themselves. Specific and higher priority to override the
/usr/share/doc rule already included.

Signed-off-by: Peter O'Connor <peter@solus-project.com>