rc: Add support for easy rubygems
More ruby is starting to enter Solus, and currently they are all bundling their
own gems. This means if two packages use the same gem they conflict. This makes
it easy to package Rubygems so with start doing that instead (and it becomes
like perl/python with lots of little deps). It should also allow for networking
to be turned off for these packages. See T5014

It uses * to find the files as it will differ from the package name in many
cases (particularly where you need to add ruby-, but you can't remove it as
that breaks packages who's name uses ruby-). There should only be one per
package, but may be in different locations.

It may need some tweaks but will allow someone to start fixing up our ruby!

Signed-off-by: Peter O'Connor peter@solus-project.com