context: Drop -f{function,data}-sections, fix invalid filtering
We drop these 2 cflags as experiments have shown these are only useful
when producing a library without LTO that you then wish to link with using
LTO. The gains are minimal in this area and by default leads to much
larger objects.

Additionally we fix the optimize_flags function to only strip existing
optimisation levels when dealing with the speed/size keys. This ensures we
don't strip the -O3 or -Os when adding a second `lto` line or similar.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>