context: Only use `-Os` for GCC, clang breaks `-Os`
Long standing issue upstream whereby -flto -Os do not work together as the
gold linker refuses anything outside the range of -O0 to -O3. To mitigate
this and other potential issues in the future, we'll force -O2 when using
clang and the size flag, as there is nothing else we can realistically do.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>