Define two new helper macros
The PKG_ROOT_DIR will point to the root of the packages tree, which is
always one level HIGHER than the build directory.

The PKG_BUILD_DIR will point to the current build directory for the
build operation, i.e. `build`, `build-32`, which is always one level HIGHER
than the workdir.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>