context: Teach it all about automatic job counts
If we have 'auto' set in the eopkg.conf file, we'll default to finding
the maximum number of jobs available. This will ensure all users have
a sane build config set without having to edit eopkg.conf anymore.

This differs slightly to the current eopkg auto implementation of jobs+1
which is a bit taxing for most systems, so will be corrected.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>