Support prefixless package names
This is very much a corner case for when we explicitly need differently
named packages, as we prefer logical organisation with a common prefix.

However, this option allows one to prefix their pattern (package key) with
'^' to set an explicit name, that doesn't inherit a prefix from the main
source package name. This will be useful for a small handful of cases like
the complex GTK3 migration to ypkg.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>