Implement automatic kernel module dependencies
Any time we encounter a .ko file we'll run modinfo on it to determine
exactly which version it detects on. We then rely on a Solusism to build
the System.map file location and find the owning package for that given
kernel version.

This gives us automatic kernel dependencies with a minimum release field,
though we'll likely set absolute release requirements next. If we happen
to encounter a System.map file ourselves, we'll record that as providing
a kernel, so that we don't self depend in our kernel builds. Also this
allows us to split up kernel packages if we really need to.

In future this location will be more readily expressed in the index so
that we can expose kernel capabilities directly.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>