meta: Stop overwriting utime/mtime of our files
This is now having an effect on packages where we need to see the mtime
change for triggers to work properly, such as for usysconf. Given that
this doesn't affect deltas, and only reproducibility, and that our
deltas kinda suck anyway, it's no big loss.

We'll revisit this when we replace eopkg with sol.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>