dependencies: Work around ugly bug in eopkg with filesdb
When updating and having files conflicts by swapping some files around,
eopkg filesdb becomes distinctly less useful by "forgetting" the keys in
the filesdb and can only be remedied by an `eopkg rdb` call.

In those instances we'll fallback to a slow search and try to find the
guy in question.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>